+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased][unreleased]
 ### Added
 - Arabic localication.
+- Sending files from other apps [#182](https://github.com/Antidote-for-Tox/Antidote/issues/182).
 - Highlighting links in chat [#176](https://github.com/Antidote-for-Tox/Antidote/issues/176).
 
 ### Changed